split files: Moved colour conversions to colour.py
Issue: #22
